+++ src/interface/mapping/endpoint.rs
@@ -23,7 +23,7 @@ use std::{
 };
 
 use itertools::{EitherOrBoth, Itertools};
-use log::{debug, info, trace};
+use log::{debug, error, info, trace};
 
 use super::path::MappingPath;
 
@@ -41,14 +41,29 @@ use super::path::MappingPath;
 /// For more information see [Astarte - Docs](https://docs.astarte-platform.org/astarte/latest/030-interface.html#limitations)
 ///
 /// The endpoints uses Cow to not allocate the string if an error occurs.
-///
 #[derive(Debug, PartialEq, Eq, Clone)]
-pub(crate) struct Endpoint<T> {
+pub struct Endpoint<T> {
     pub(super) path: T,
     pub(super) levels: Vec<Level<T>>,
 }
 
 impl<T> Endpoint<T> {
+    /// Check that the endpoint is equal to the given mapping
+    pub fn eq_mapping<S>(&self, mapping: S) -> bool
+    where
+        S: AsRef<str>,
+        T: AsRef<str>,
+    {
+        match MappingPath::try_from(mapping.as_ref()) {
+            Ok(mapping) => mapping.eq(self),
+            Err(err) => {
+                error!("failed to parse mapping {err}");
+
+                false
+            }
+        }
+    }
+
     /// Iter the levels of the endpoint.
     pub(crate) fn iter(&self) -> SliceIter<Level<T>> {
         self.levels.iter()