+++ astarte-device-sdk-derive/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Macro to implement the `FromEvent` trait on a generic struct.
+
 ### Changed
 - Update the `AstarteAggregate` derive macro to syn `2`, see
   [#236](https://github.com/astarte-platform/astarte-device-sdk-rust/pull/236).